refactor: best practice according to review change-requestes
