+++ src/components/__test__/PageIntro.test.js
@@ -1,6 +1,6 @@
 import renderer from "react-test-renderer";
 
-import PageIntro from "../page-intro";
+import PageIntro from "../PageIntro";
 
 it("renders correctly", () => {
     const tree = renderer.create(<PageIntro />).toJSON();
